library manegement with c programming language
Library Project - Actions to be Taken
Enum Structure to Define:
1. Levels of Authority: student->0, will have employee-> value.
Structs to be Defined:
1. personinfo: Character string type variables in the name of userName, memberSurname,
IDlibrary, integer type variables with password name and AuthorityLevels enum
will contain a variable of type .
2. Book Information: Name,kAuthorName,kAuthorSurname,character strings in the name of the publisher
Variables of type, PageNumber, integer variables named BookStatus
will include.
What to Do in the Application:
1. All member information will be stored in an array of type memberinfo named memberList.
2. All book information will be stored in an array of bookinfo type named bookList.
3. All book and member information will be taken from files with .txt extension and will be in series.
will be appointed.
4. The user will be provided with a registration or login process.
5. If the entrant is a student:
a. Book viewing
b. book search
6. If the person entering is an employee;
a. Adding a new book
b. View members
Application Flow:
First, the member and book information stored in the .txt file is read from the file.
will be transferred to arrays and all information will be stored in these arrays. Then, the user
registration or login will be required. Student of the person logging into the system
or 2 different menus will be presented depending on whether it is an employee or not. In this menu
All the above-mentioned operations must be performed, the user 'Q'
The menu will appear continuously as long as you don't press it.
Functions and Duties to be Defined:
Query 1st person(personinfo *memberList,int ID): This function returns the memberList and ID value
will take. If this ID value has been used before, it will return the ID value,
which will return -1 if not used.
2. register(personinfo *memberList): This function will retrieve the memberList array.
A new record will be made in the last index of the userList array. Registration process
A random ID value will be determined during
If not used, the ID value will be assigned, otherwise the new ID value will be determined.
Then, the registration process with information such as name, surname, student-employee, password.
will be completed.
3. View books(bookinfo *library): This function will display all book information
will get the library array where it is stored and all the books; book title-book author name,
the name of the book publisher, the number of pages in the book and the availability of the book.
will print.
4.Search for book(bookinfo*library): This function is where all book information is stored.
will take the library array and perform a search based on the book the user wants to search for.
will be done. If there is a book sought, the information will be printed, otherwise the book will be
It will be said that it is not.
5.Add book(bookinfo *library): This function is where all book information is stored.
will take the library array and the name of the book to be added, the author's name of the book,
will be added to the library series with information such as the publisher's name, the number of pages of the book.
6. MembersView(personinfo *memberList): This function displays all member information
will take the memberList array where it is stored and print all the members' information.
7. Login(personinfo*memberList): This function is where all member information is stored.
It will take the memberList sequence and login process will be provided according to the ID, password information.
If both the ID and password information are correct, the login process will be performed and
authority will be assigned (Student-Staff)
8. Import members(personinfo *memberList): This function returns all member information
will get the memberList array where it will be stored and all members' information will be extracted from the file.
will be read and assigned to the memberList array.
9. Transfer books(bookinfo *library): This function allows you to export all book information.
will get the library array where it will be stored and all book information will be read from the file
will be assigned to the library array.
10.Backup members(personinfo *memberList): This function returns all member information
will get the array of membersList where it is stored and all members' information will be read from the array.
will be saved in a formatted file.
11.Transfer books(bookinfo *library): This function allows you to export all book information.
will get the library array where it is stored and all book information will be read from the array and put into the file
will be saved in the format